+++ dpp_scripts/yaml_helper.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 import psrqpy
 import logging
 import os
