+++ mwa_search/obs_tools.py
@@ -1,4 +1,7 @@
-from astropy.coordinates import SkyCoord
+from mwa_pb.mwa_tile import h2e
+
+from astropy.coordinates import SkyCoord, EarthLocation, AltAz
+from astropy.time import Time
 import astropy.units as u
 
 # vcstools imports
@@ -6,10 +9,82 @@ import checks
 from config_vcs import load_config_file
 
 # mwa_search imports
-from grid import get_grid
+from mwa_search.grid_tools import get_grid
 
 import logging
 logger = logging.getLogger(__name__)
+
+
+def getTargetAZZA(ra,dec,time,lat=-26.7033,lon=116.671,height=377.827):
+    """
+    Function to get the target position in alt/az at a given EarthLocation and Time.
+
+    Default lat,lon,height is the centre of  MWA.
+
+    Input:
+      ra - target right ascension in astropy-readable format
+      dec - target declination in astropy-readable format
+      time - time of observation in UTC (i.e. a string on form: yyyy-mm-dd hh:mm:ss.ssss)
+      lat - observatory latitude in degrees
+      lon - observatory longitude in degrees
+
+    Returns:
+      a list containing four elements in the following order:
+        list[0] = target azimuth in radians
+        list[1] = target zenith angle in radians
+        list[2] = target azimuth in degrees
+        list[3] = target zenith angle in degrees
+    """
+    #print "Creating EarthLocation for: lat = {0} deg, lon = {1} deg".format(lat,lon)
+    location = EarthLocation(lat=lat*u.deg, lon=lon*u.deg, height=height*u.m)
+
+    #print "Creating SkyCoord for target at (RA,DEC) = ({0},{1})".format(ra,dec)
+    coord = SkyCoord(ra,dec,unit=(u.hourangle,u.deg))
+    #print "Target at: ({0},{1}) deg".format(coord.ra.deg,coord.dec.deg)
+
+    obstime = Time(time)
+    #print "Observation time: {0}".format(obstime.iso)
+
+    #print "Converting to Alt,Az for given time and observatory location..."
+    altaz = coord.transform_to(AltAz(obstime=obstime,location=location))
+    #print "Target (Alt,Az) = ({0},{1}) deg".format(altaz.alt.deg,altaz.az.deg)
+
+    #print "Converting to (Az,ZA)"
+    az = altaz.az.rad
+    azdeg = altaz.az.deg
+
+    za = np.pi/2 - altaz.alt.rad
+    zadeg = 90 - altaz.alt.deg
+
+    #print "Target (Az,ZA) = ({0},{1}) deg".format(azdeg,zadeg)
+
+    return [az,za,azdeg,zadeg]
+
+
+def getTargetradec(az,za,time,lst,lat=-26.7033,lon=116.671,height=377.827):
+    """
+    Function to get the target position in ra dec at a given EarthLocation and Time.
+
+    Default lat,lon,height is the centre of  MWA.
+
+    Input:
+      az - target aximuth in radians
+      za - target zenith in radians
+      time - time of observation in UTC (i.e. a string on form: yyyy-mm-dd hh:mm:ss.ssss)
+      lat - observatory latitude in degrees
+      lon - observatory longitude in degrees
+
+    Returns:
+      a list containing four elements in the following order:
+        list[0] = target ra in degrees
+        list[1] = target dec in degrees
+    """
+
+    ha,dec = h2e(az,za,lat) #hour angle and dec in degrees
+    ra = lst-ha
+
+
+    return [ra,dec]
 
 def check_data(obsid, beg=None, dur=None, base_dir=None):
     """
