+++ scripts/grid.py
@@ -1,363 +1,25 @@
 #!/usr/bin/env python3
-import find_pulsar_in_obs as fpio
-
-import matplotlib.pyplot as plt
-from matplotlib import patches
-
-# numerical and maths modules
-import numpy as np
+import argparse
 from astropy.coordinates import SkyCoord,EarthLocation,AltAz
 from astropy.time import Time
 import astropy.units as u
-from math import cos,sin
+from math import cos
+import numpy as np
 
-#from mpi4py import MPI
-import argparse
+from matplotlib import use
+use('Agg')
+import matplotlib.pyplot as plt
+from matplotlib import patches
+
 from mwa_pb.mwa_tile import h2e
-#import urllib
-#import urllib2
-#import json
 
+# vcstools imports
 import mwa_metadb_utils as meta
+import find_pulsar_in_obs as fpio
 
-
-def getTargetAZZA(ra,dec,time,lat=-26.7033,lon=116.671,height=377.827):
-    """
-    Function to get the target position in alt/az at a given EarthLocation and Time.
-
-    Default lat,lon,height is the centre of  MWA.
-
-    Input:
-      ra - target right ascension in astropy-readable format
-      dec - target declination in astropy-readable format
-      time - time of observation in UTC (i.e. a string on form: yyyy-mm-dd hh:mm:ss.ssss)
-      lat - observatory latitude in degrees
-      lon - observatory longitude in degrees
-
-    Returns:
-      a list containing four elements in the following order:
-        list[0] = target azimuth in radians
-        list[1] = target zenith angle in radians
-        list[2] = target azimuth in degrees
-        list[3] = target zenith angle in degrees
-    """
-    #print "Creating EarthLocation for: lat = {0} deg, lon = {1} deg".format(lat,lon)
-    location = EarthLocation(lat=lat*u.deg, lon=lon*u.deg, height=height*u.m)
-
-    #print "Creating SkyCoord for target at (RA,DEC) = ({0},{1})".format(ra,dec)
-    coord = SkyCoord(ra,dec,unit=(u.hourangle,u.deg))
-    #print "Target at: ({0},{1}) deg".format(coord.ra.deg,coord.dec.deg)
-
-    obstime = Time(time)
-    #print "Observation time: {0}".format(obstime.iso)
-
-    #print "Converting to Alt,Az for given time and observatory location..."
-    altaz = coord.transform_to(AltAz(obstime=obstime,location=location))
-    #print "Target (Alt,Az) = ({0},{1}) deg".format(altaz.alt.deg,altaz.az.deg)
-
-    #print "Converting to (Az,ZA)"
-    az = altaz.az.rad
-    azdeg = altaz.az.deg
-
-    za = np.pi/2 - altaz.alt.rad
-    zadeg = 90 - altaz.alt.deg
-
-    #print "Target (Az,ZA) = ({0},{1}) deg".format(azdeg,zadeg)
-
-    return [az,za,azdeg,zadeg]
-
-
-def getTargetradec(az,za,time,lst,lat=-26.7033,lon=116.671,height=377.827):
-    """
-    Function to get the target position in ra dec at a given EarthLocation and Time.
-
-    Default lat,lon,height is the centre of  MWA.
-
-    Input:
-      az - target aximuth in radians
-      za - target zenith in radians
-      time - time of observation in UTC (i.e. a string on form: yyyy-mm-dd hh:mm:ss.ssss)
-      lat - observatory latitude in degrees
-      lon - observatory longitude in degrees
-
-    Returns:
-      a list containing four elements in the following order:
-        list[0] = target ra in degrees
-        list[1] = target dec in degrees
-    """
-
-    ha,dec = h2e(az,za,lat) #hour angle and dec in degrees
-    ra = lst-ha
-
-
-    return [ra,dec]
-
-
-def two_floats(value):
-    values = value.split()
-    if len(values) != 2:
-        raise argparse.ArgumentError
-    return values
-
-
-#gird movements all in rad
-def left(ra_in, dec_in, fwhm):
-    dec_out = dec_in
-    ra_out = ra_in - fwhm/cos(dec_in)
-    return [ra_out,dec_out]
-
-def right(ra_in, dec_in, fwhm):
-    dec_out = dec_in
-    ra_out = ra_in + fwhm/cos(dec_in)
-    return [ra_out,dec_out]
-
-def up(ra_in, dec_in, fwhm):
-    dec_out = dec_in + fwhm / cos(dec_in + np.radians(26.7))**2
-    ra_out = ra_in
-    return [ra_out,dec_out]
-
-def down(ra_in, dec_in, fwhm):
-    dec_out = dec_in - fwhm / cos(dec_in + np.radians(26.7))**2
-    ra_out = ra_in
-    return [ra_out,dec_out]
-
-def up_left(ra_in, dec_in, fwhm):
-    half_fwhm_approx = fwhm/2.#/cos(dec_in)
-    dec_out = dec_in + sin(np.radians(60.))*sin(half_fwhm_approx) / sin(np.radians(30.)) \
-              / cos(dec_in + np.radians(26.7))**2
-    ra_out = ra_in - fwhm/2./cos(dec_out)
-    return [ra_out,dec_out]
-
-def up_right(ra_in, dec_in, fwhm):
-    half_fwhm_approx = fwhm/2.#/cos(dec_in)
-    dec_out = dec_in + sin(np.radians(60.))*sin(half_fwhm_approx) / sin(np.radians(30.)) \
-              / cos(dec_in + np.radians(26.7))**2
-    ra_out = ra_in + fwhm/2./cos(dec_out)
-    return [ra_out,dec_out]
-
-def down_left(ra_in, dec_in, fwhm):
-    half_fwhm_approx = fwhm/2.#/cos(dec_in)
-    dec_out = dec_in - sin(np.radians(60.))*sin(half_fwhm_approx) / sin(np.radians(30.)) \
-              / cos(dec_in + np.radians(26.7))**2
-    ra_out = ra_in - fwhm/2./cos(dec_out)
-    return [ra_out,dec_out]
-
-def down_right(ra_in, dec_in, fwhm):
-    half_fwhm_approx = fwhm/2.#/cos(dec_in)
-    dec_out = dec_in - sin(np.radians(60.))*sin(half_fwhm_approx) / sin(np.radians(30.)) \
-              / cos(dec_in + np.radians(26.7))**2
-    ra_out = ra_in + fwhm/2./cos(dec_out)
-    return [ra_out,dec_out]
-
-
-def cross_grid(ra0,dec0,centre_fwhm, loop):
-    #start location list [loop number][shape corner (6 for hexagon 4 for square)][number from corner]
-    #each item has [ra,dec,fwhm] in radians
-    pointing_list = [[[[ra0,dec0,centre_fwhm]]]]
-    print("Calculating the tile positions")
-    for l in range(loop):
-        loop_temp = []
-        for c in range(4):
-            if c == 0:
-                ra,dec =left(pointing_list[l][c][0][0],
-                             pointing_list[l][c][0][1],centre_fwhm)
-            elif c == 1:
-                if l == 0:
-                    c = 0
-                ra,dec =up(pointing_list[l][c][0][0],
-                           pointing_list[l][c][0][1],centre_fwhm)
-            elif c == 2:
-                if l == 0:
-                    c = 0
-                ra,dec =right(pointing_list[l][c][0][0],
-                              pointing_list[l][c][0][1],centre_fwhm)
-            elif c == 3:
-                if l == 0:
-                    c = 0
-                ra,dec =down(pointing_list[l][c][0][0],
-                             pointing_list[l][c][0][1],centre_fwhm)
-            loop_temp.append([[ra, dec]])
-        pointing_list.append(loop_temp)
-    return pointing_list
-
-
-def hex_grid(ra0,dec0,centre_fwhm, loop):
-    #start location list [loop number][shape corner (6 for hexagon 4 for square)][number from corner]
-    #each item has [ra,dec,fwhm] in radians
-    pointing_list = [[[[ra0,dec0,centre_fwhm]]]]
-    print("Calculating the tile positions")
-
-    for l in range(loop):
-        #different step for each corner
-        loop_temp = []
-        for c in range(6):
-            corner_temp = []
-            for n in range(l + 1):
-                if l == 0:
-                    #First loop so all c = 0
-                    if c == 0:
-                        ra,dec =left(pointing_list[l][0][n][0],
-                                     pointing_list[l][0][n][1],centre_fwhm)
-                    elif c == 1:
-
-                        ra,dec =up_left(pointing_list[l][0][n][0],
-                                        pointing_list[l][0][n][1],centre_fwhm)
-                    elif c == 2:
-                        ra,dec =up_right(pointing_list[l][0][n][0],
-                                         pointing_list[l][0][n][1],centre_fwhm)
-                    elif c == 3:
-                        ra,dec =right(pointing_list[l][0][n][0],
-                                      pointing_list[l][0][n][1],centre_fwhm)
-                    elif c == 4:
-                        ra,dec =down_right(pointing_list[l][0][n][0],
-                                           pointing_list[l][0][n][1],centre_fwhm)
-                    elif c == 5:
-                        ra,dec =down_left(pointing_list[l][0][n][0],
-                                          pointing_list[l][0][n][1],centre_fwhm)
-
-                elif l == n:
-                    #change the 2 for each loop
-                    #uses next corner
-                    if c == 0:
-                        ra,dec =left(pointing_list[l][c+1][0][0],
-                                     pointing_list[l][c+1][0][1],centre_fwhm)
-                    elif c == 1:
-                        ra,dec =up_left(pointing_list[l][c+1][0][0],
-                                     pointing_list[l][c+1][0][1],centre_fwhm)
-                    elif c == 2:
-                        ra,dec =up_right(pointing_list[l][c+1][0][0],
-                                     pointing_list[l][c+1][0][1],centre_fwhm)
-                    elif c == 3:
-                        ra,dec =right(pointing_list[l][c+1][0][0],
-                                     pointing_list[l][c+1][0][1],centre_fwhm)
-                    elif c == 4:
-                        ra,dec =down_right(pointing_list[l][c+1][0][0],
-                                     pointing_list[l][c+1][0][1],centre_fwhm)
-                    elif c == 5:
-                        ra,dec =down_left(pointing_list[l][0][0][0],
-                                 pointing_list[l][0][0][1],centre_fwhm)
-
-                else:
-                    if c == 0:
-                        ra,dec =left(pointing_list[l][c][n][0],
-                                     pointing_list[l][c][n][1],centre_fwhm)
-                    elif c == 1:
-
-                        ra,dec =up_left(pointing_list[l][c][n][0],
-                                        pointing_list[l][c][n][1],centre_fwhm)
-                    elif c == 2:
-                        ra,dec =up_right(pointing_list[l][c][n][0],
-                                         pointing_list[l][c][n][1],centre_fwhm)
-                    elif c == 3:
-                        ra,dec =right(pointing_list[l][c][n][0],
-                                      pointing_list[l][c][n][1],centre_fwhm)
-                    elif c == 4:
-                        ra,dec =down_right(pointing_list[l][c][n][0],
-                                           pointing_list[l][c][n][1],centre_fwhm)
-                    elif c == 5:
-                        ra,dec =down_left(pointing_list[l][c][n][0],
-                                          pointing_list[l][c][n][1],centre_fwhm)
-                corner_temp.append([ra,dec])
-            loop_temp.append(corner_temp)
-        pointing_list.append(loop_temp)
-    return pointing_list
-
-
-def square_grid(ra0,dec0,centre_fwhm, loop):
-    #start location list [loop number][shape corner (4 for square)][number from corner]
-    #each item has [ra,dec,fwhm] in radians
-    pointing_list = [[[[ra0,dec0,centre_fwhm]]]]
-    print("Calculating the tile positions")
-
-    for l in range(loop):
-        #different step for each corner
-        loop_temp = []
-        for c in range(4):
-            corner_temp = []
-            for n in range((l + 1) * 2):
-                if n == 0:
-                    #grab from previous corner
-                    cfrom = (c + 3)%4
-                    if l == 0:
-                        #if first loop set all corners to zero
-                        cfrom = 0
-                    #and last number from corner
-                    nfrom = l * 2 - 1
-                    #First loop so all c = 0
-                    if c == 0:
-                        ra,dec = left(pointing_list[l][cfrom][nfrom][0],
-                                      pointing_list[l][cfrom][nfrom][1],centre_fwhm)
-                    elif c == 1:
-                        ra,dec = up(pointing_list[l][cfrom][nfrom][0],
-                                    pointing_list[l][cfrom][nfrom][1],centre_fwhm)
-                    elif c == 2:
-                        ra,dec = right(pointing_list[l][cfrom][nfrom][0],
-                                       pointing_list[l][cfrom][nfrom][1],centre_fwhm)
-                    elif c == 3:
-                        ra,dec = down(pointing_list[l][cfrom][nfrom][0],
-                                      pointing_list[l][cfrom][nfrom][1],centre_fwhm)
-
-                else:
-                    #moves to the edges
-                    if c == 0:
-                        ra,dec = up(corner_temp[n-1][0],
-                                    corner_temp[n-1][1],centre_fwhm)
-                    elif c == 1:
-                        ra,dec = right(corner_temp[n-1][0],
-                                       corner_temp[n-1][1],centre_fwhm)
-                    elif c == 2:
-                        ra,dec = down(corner_temp[n-1][0],
-                                      corner_temp[n-1][1],centre_fwhm)
-                    elif c == 3:
-                        ra,dec = left(corner_temp[n-1][0],
-                                      corner_temp[n-1][1],centre_fwhm)
-
-
-                corner_temp.append([ra,dec])
-            loop_temp.append(corner_temp)
-        pointing_list.append(loop_temp)
-    return pointing_list
-
-
-def get_grid(ra, dec, grid_sep, loop, grid_type='hex'):
-    """
-    ra: Right Acension in radians
-    dec: Declination in radians
-    grid_sep: seperation between grid pointings in radians
-    loop: number of pointing loops
-    grid_type: Possible grid types from ['hex', 'cross', 'squaire']
-
-    return [rads, decds]
-    RAs and Decs in degrees
-    """
-    #calc grid positions
-    if grid_type == 'hex':
-        pointing_list = hex_grid(   ra, dec, grid_sep, loop)
-    elif grid_type == 'cross':
-        pointing_list = cross_grid( ra, dec, grid_sep, loop)
-    elif grid_type == 'square':
-        pointing_list = square_grid(ra, dec, grid_sep, loop)
-    else:
-        print("Unrecognised grid type. Exiting.")
-        quit()
-    #TODO add square
-
-    rads = []; decds = []
-
-    print("Converting ra dec to degrees")
-    for loop in pointing_list:
-        for corner in loop:
-            for num in corner:
-                #format grid pointings
-                rad = np.degrees(num[0])
-                decd = np.degrees(num[1])
-
-                if (decd < 90.) and (decd > -90.):
-                    # Only include ra and dec within the real decs
-                    rads.append(rad)
-                    decds.append(decd)
-    return rads, decds
-
+# mwa_search imports
+from mwa_search.obs_tools import getTargetAZZA
+from mwa_search.grid_tools import get_grid
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description="""
