+++ setup.py
@@ -52,22 +52,22 @@ setup(name="mwa_search",
       url="https://github.com/NickSwainston/mwa_search",
       #long_description=read('README.md'),
       python_requires='>=3.6',
-      packages=['dpp'],
+      packages=['dpp', 'mwa_search'],
       install_requires=reqs,
       scripts=['scripts/check_known_pulsars.py', 'scripts/cold_storage_mover.py',
                'scripts/grid.py', 'scripts/lfDDplan.py',
-               'scripts/mwa_search_pipeline.py', 'scripts/pulsars_in_fov.py',
+               'scripts/mwa_search_pipeline.py',
                'scripts/splice_wrapper.py', 'scripts/LOTAAS_wrapper.py',
                'scripts/search_launch_loop.sh', 'scripts/rsync_rm_loop.sh',
                'scripts/bestgridpos.py', 'scripts/data_processing_pipeline.py',
+               'dpp_scripts/stokes_fold.py', 'dpp_scripts/yaml_helper.py',
+               'dpp_scripts/pulsars_in_fov.py',
                'plotting/plot_obs_pulsar.py',
                'plotting/position_sn_heatmap_fwhm.py',
-               'dpp/binfinder.py', 'dpp/stokes_fold.py', 'dpp/plotting_toolkit.py',
-               'dpp/yaml_helper.py', 'dpp/prepfold_cmd_make.py',
-               'dpp/pulsar_obs_helper.py', 'dpp/dpp_check_args.py', 'dpp/misc_helper.py', 'dpp/upload_and_move.py',
+               'dpp/plotting_toolkit.py',
                'version.py',
                'nextflow/beamform.nf', 'nextflow/beamform_module.nf',
-                'nextflow/nextflow.config',
+               'nextflow/nextflow.config',
                'nextflow/candidate_TOAs.nf', 'nextflow/find_candidate_position.nf',
                'nextflow/classifier.nf', 'nextflow/classifier_module.nf',
                'nextflow/pdmp.nf', 'nextflow/dspsr_module.nf',
